Deleted dead code
